remove patch.py for CI
