+++ tests/a_scripts/data_prep/test_convert_delta_to_json.py
@@ -4,7 +4,6 @@
 # copyright 2022 mosaicml llm foundry authors
 # spdx-license-identifier: apache-2.0
 
-import unittest
 from argparse import Namespace
 from typing import Any
 from unittest.mock import MagicMock, patch
@@ -54,7 +53,3 @@ class TestStreamDeltaToJson():
         mock_to_json.assert_called()
         mock_cursor.close.assert_called()
         mock_connection.close.assert_called()
-
-
-if __name__ == '__main__':
-    unittest.main()
